Vinyl record exchanger in C#
This version collects user credentials and writes them to a text file. This is really a backup of the currentstate. My instructor recommended trying to use JSON instead of text files. I have done a brief research, and I really want to try it. I still need to complete the project by the end of this week so I might roll back to the text-file approach for data exchange.